prevent users from changing/deleting travels not their own

+++ controller/travels.js
@@ -24,21 +24,27 @@ class TravelsController {
 
     async updateTravel(req, res) {
         let travel_id = req.params.travel_id
+        let user_id = req.session.user_id
         try {
-            for (const key in req.body) {
-                if (!['name', 'start', 'end', 'destination'].includes(key)) {
-                    continue
+            let travel_ids = await travelsService.getTravelIDsOfUser(user_id)
+            if (travel_ids.includes(travel_id)) {
+                for (const key in req.body) {
+                    if (!['name', 'start', 'end', 'destination'].includes(key)) {
+                        continue
+                    }
+                    await travelsService.updateTravel(travel_id, key, req.body[key])
                 }
-                await travelsService.updateTravel(travel_id, key, req.body[key])
+            } else {
+                res.status(401).json('Unauthorized')
             }
         } catch (e) {
             res.status(400).json('Bad request')
         }
     }
 
-    async getTravels(req, res) {
+    async getAllUserTravels(req, res) {
         try {
-            const travels = await travelsService.getTravels(req.session.user_id)
+            const travels = await travelsService.getAllUserTravels(req.session.user_id)
             res.status(200).json(travels)
         } catch (e) {
             res.status(400).json('Bad request')
@@ -47,9 +53,15 @@ class TravelsController {
 
     async deleteTravel(req, res) {
         let travel_id = req.params.travel_id
+        let user_id = req.session.user_id
         try {
-            await travelsService.deleteTravel(travel_id)
-            res.status(200).json('OK')
+            let travel_ids = await travelsService.getTravelIDsOfUser(user_id)
+            if (travel_ids.includes(travel_id)) {
+                await travelsService.deleteTravel(travel_id)
+                res.status(200).json('OK')
+            } else {
+                res.status(401).json('Unauthorized')
+            }
         } catch (e) {
             res.status(400).json('Bad request')
         }
